move app init to dop
