+++ modules/core-services/services/application/appliation.go
@@ -19,22 +19,17 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"net/url"
 	"strconv"
-	"strings"
 	"time"
 
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
-	"gopkg.in/yaml.v2"
 
 	cmspb "github.com/erda-project/erda-proto-go/core/pipeline/cms/pb"
 	"github.com/erda-project/erda/apistructs"
 	"github.com/erda-project/erda/bundle"
-	"github.com/erda-project/erda/modules/core-services/conf"
 	"github.com/erda-project/erda/modules/core-services/dao"
 	"github.com/erda-project/erda/modules/core-services/model"
-	"github.com/erda-project/erda/modules/core-services/services/apierrors"
 	"github.com/erda-project/erda/modules/core-services/types"
 	"github.com/erda-project/erda/pkg/common/apis"
 	"github.com/erda-project/erda/pkg/crypto/uuid"
@@ -300,121 +295,6 @@ func (a *Application) UpdateWithEvent(appID int64, updateReq *apistructs.Applica
 	}
 
 	return app, nil
-}
-
-// Init 应用初始化
-func (a *Application) Init(initReq *apistructs.ApplicationInitRequest) (uint64, error) {
-	// 1. 参数校验
-	// 2. 生成 pipeline.yml
-	// 3. 调用 pipeline API 创建 pipeline
-	if initReq.MobileAppName == "" {
-		return 0, apierrors.ErrInitApplication.MissingParameter("mobileAppName")
-	}
-	if initReq.BundleID == "" {
-		return 0, apierrors.ErrInitApplication.MissingParameter("bundleID")
-	}
-	if initReq.PackageName == "" {
-		return 0, apierrors.ErrInitApplication.MissingParameter("packageName")
-	}
-	if initReq.MobileDisplayName == "" {
-		initReq.MobileDisplayName = initReq.MobileAppName
-	}
-
-	app, err := a.db.GetApplicationByID(int64(initReq.ApplicationID))
-	if err != nil {
-		return 0, err
-	}
-	if app.Mode != string(apistructs.ApplicationModeMobile) {
-		return 0, errors.Errorf("only support mobile app template init")
-	}
-
-	actionProjectName := strings.Replace(initReq.MobileAppName, "-", "", -1)
-	// generate mobile template action
-	mobileTemplateStage := &apistructs.PipelineYmlAction{
-		Type: "mobile-template",
-		Params: map[string]interface{}{
-			"project_name": actionProjectName,
-			"display_name": initReq.MobileDisplayName,
-			"bundle_id":    initReq.BundleID,
-			"package_name": initReq.PackageName,
-		},
-	}
-
-	// generate remote url
-	var token string
-	if members, err := a.db.GetMemberByScopeAndUserID(initReq.UserID, apistructs.OrgScope, app.OrgID); err == nil && members != nil {
-		// TODO: kanbudong
-		token = members[0].Token
-	}
-	if token == "" {
-		return 0, errors.Errorf("not found user token")
-	}
-	org, err := a.db.GetOrg(app.OrgID)
-	if err != nil {
-		return 0, err
-	}
-	u, _ := url.Parse(conf.UIPublicURL())
-	remoteUrl := fmt.Sprintf("%s://git:%s@%s/%s/dop/%s/%s", u.Scheme, token, conf.UIDomain(), org.Name, app.ProjectName, app.Name)
-
-	// generate git push action
-	gitPushStage := &apistructs.PipelineYmlAction{
-		Type: "git-push",
-		Params: map[string]interface{}{
-			"workdir":    "${mobile-template}/" + actionProjectName,
-			"remote_url": remoteUrl,
-		},
-	}
-
-	// generate pipeline.yml
-	pys := apistructs.PipelineYml{
-		Version: "1.1",
-		Stages: [][]*apistructs.PipelineYmlAction{
-			{
-				mobileTemplateStage,
-			},
-			{
-				gitPushStage,
-			},
-		},
-	}
-
-	// fetch cluster name
-	project, err := a.db.GetProjectByID(app.ProjectID)
-	if err != nil {
-		return 0, err
-	}
-	clusterConfig := map[string]string{}
-	json.Unmarshal([]byte(project.ClusterConfig), &clusterConfig)
-	clusterName, ok := clusterConfig[string(apistructs.DevWorkspace)]
-	if !ok {
-		return 0, errors.Errorf("not found cluster")
-	}
-
-	ymlContent, err := yaml.Marshal(&pys)
-	if err != nil {
-		return 0, err
-	}
-	req := &apistructs.PipelineCreateRequestV2{
-		PipelineYml:     string(ymlContent),
-		PipelineYmlName: fmt.Sprintf("%s_%s_pipeline.yml", project.Name, app.Name),
-		PipelineSource:  apistructs.PipelineSourceDice,
-		ClusterName:     clusterName,
-		Labels: map[string]string{
-			apistructs.LabelBranch:        "master",
-			apistructs.LabelOrgID:         strconv.FormatInt(app.OrgID, 10),
-			apistructs.LabelProjectID:     strconv.FormatInt(app.ProjectID, 10),
-			apistructs.LabelAppID:         strconv.FormatInt(app.ID, 10),
-			apistructs.LabelDiceWorkspace: string(apistructs.ProdWorkspace),
-		},
-		AutoRunAtOnce: true,
-	}
-	// create pipeline info
-	pipelineInfo, err := a.bdl.CreatePipeline(req)
-	if err != nil {
-		return 0, err
-	}
-
-	return pipelineInfo.ID, nil
 }
 
 // Update 更新应用
