+++ modules/dop/initialize.go
@@ -39,6 +39,7 @@ import (
 	"github.com/erda-project/erda/modules/dop/services/apidocsvc"
 	"github.com/erda-project/erda/modules/dop/services/apierrors"
 	"github.com/erda-project/erda/modules/dop/services/appcertificate"
+	"github.com/erda-project/erda/modules/dop/services/application"
 	"github.com/erda-project/erda/modules/dop/services/assetsvc"
 	"github.com/erda-project/erda/modules/dop/services/autotest"
 	atv2 "github.com/erda-project/erda/modules/dop/services/autotest_v2"
@@ -529,6 +530,11 @@ func (p *provider) initEndpoints(db *dao.DBClient) (*endpoints.Endpoints, error)
 		project.WithCMP(p.Cmp),
 	)
 
+	app := application.New(
+		application.WithBundle(bdl.Bdl),
+		application.WithDBClient(db),
+	)
+
 	codeCvc := code_coverage.New(
 		code_coverage.WithDBClient(db),
 		code_coverage.WithBundle(bdl.Bdl),
@@ -562,6 +568,7 @@ func (p *provider) initEndpoints(db *dao.DBClient) (*endpoints.Endpoints, error)
 		endpoints.WithAssetSvc(assetsvc.New(assetsvc.WithBranchRuleSvc(branchRule))),
 		endpoints.WithFileTreeSvc(filetreeSvc),
 		endpoints.WithProject(proj),
+		endpoints.WithApplication(app),
 
 		endpoints.WithDB(db),
 		endpoints.WithTestcase(testCaseSvc),
